# swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Snippets API",
        "description": "Test description",
        "termsOfService": "https://www.google.com/policies/terms/",
        "contact": {
            "email": "contact@snippets.local"
        },
        "license": {
            "name": "BSD License"
        },
        "version": "v1"
    },
    "basePath": "/",
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "securityDefinitions": {
        "Basic": {
            "type": "basic"
        }
    },
    "security": [
        {
            "Basic": []
        }
    ],
    "paths": {
...
        "/posts/create/": {
            "post": {
                "operationId": "Create a post",
                "description": "Create a post, post has 2 type",
                "parameters": [
                    {
                        "name": "files",
                        "in": "formData",
                        "description": "Images that's pinned to Post",
                        "required": false,
                        "type": "file"
                    },
                    {
                        "name": "content",
                        "in": "formData",
                        "description": "Content of the post",
                        "type": "string",
                        "requried": true
                    },
                    {
                        "name": "group_id",
                        "in": "formData",
                        "description": "Id of group that the post belong to",
                        "required": true,
                        "type": "number"
                    },
                    {
                        "name": "polls",
                        "in": "formData",
                        "description": "A list of Poll content you want to post",
                        "required": false,
                        "type": "object"
                    },
                    {
                        "name": "type",
                        "in": "formData",
                        "description": "Type of Post: '0' is normal post, '1' is polls with post",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response Description",
                        "schema": {
                            "$ref": "#/definitions/Post"
                        }
                    }
                },
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "posts"
                ]
            },
            "parameters": []
        },
...
        "/users/create/": {
            "get": {
                "operationId": "users_create_list",
                "description": "",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "users"
                ]
            },
            "post": {
                "operationId": "Create User",
                "description": "Create a basic user",
                "parameters": [
                    {
                        "name": "username",
                        "in": "formData",
                        "description": "Username",
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "Password",
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response Description",
                        "schema": {
                            "$ref": "#/definitions/CustomMember"
                        }
                    }
                },
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "users"
                ]
            },
            "parameters": []
        }
    },
    "definitions": {
        "CustomMember": {
            "required": [
                "username"
            ],
            "type": "object",
            "properties": {
                "avatar": {
                    "title": "Avatar",
                    "type": "string",
                    "readOnly": true,
                    "format": "uri"
                },
                "url": {
                    "title": "Url",
                    "type": "string",
                    "format": "uri",
                    "readOnly": true
                },
                "id": {
                    "title": "ID",
                    "type": "integer",
                    "readOnly": true
                },
                "username": {
                    "title": "Username",
                    "description": "Required. 150 characters or fewer. Letters, digits and @/./+/-/_ only.",
                    "type": "string",
                    "pattern": "^[\\w.@+-]+$",
                    "maxLength": 150,
                    "minLength": 1
                },
                "displayName": {
                    "title": "Display name",
                    "type": "string",
                    "maxLength": 100,
                    "x-nullable": true
                },
                "phoneNumber": {
                    "title": "Phone number",
                    "type": "string",
                    "maxLength": 15,
                    "x-nullable": true
                },
                "facebook": {
                    "title": "Facebook",
                    "type": "string",
                    "maxLength": 50,
                    "x-nullable": true
                },
                "role": {
                    "title": "Role",
                    "type": "integer",
                    "maximum": 32767,
                    "minimum": -32768,
                    "x-nullable": true
                },
                "dateOfBirth": {
                    "title": "Date of birth",
                    "type": "string",
                    "format": "date-time",
                    "x-nullable": true
                },
                "description": {
                    "title": "Description",
                    "type": "string",
                    "maxLength": 1024,
                    "x-nullable": true
                },
                "email": {
                    "title": "Email",
                    "type": "string",
                    "maxLength": 50,
                    "x-nullable": true
                },
                "participatingGroup": {
                    "title": "Participating group",
                    "type": "string",
                    "readOnly": true
                },
                "userGender": {
                    "title": "User gender",
                    "type": "integer",
                    "enum": [
                        1,
                        0,
                        2
                    ]
                },
                "cover": {
                    "title": "Cover",
                    "type": "string",
                    "readOnly": true,
                    "format": "uri"
                },
                "className": {
                    "title": "Class name",
                    "type": "string",
                    "maxLength": 15,
                    "x-nullable": true
                }
            }
        },
...
        "AssignedUserSummary": {
            "title": "Assigned user",
            "type": "object",
            "properties": {
                "id": {
                    "title": "ID",
                    "type": "integer",
                    "readOnly": true
                },
                "avatar": {
                    "title": "Avatar",
                    "type": "string",
                    "readOnly": true,
                    "format": "uri"
                },
                "displayName": {
                    "title": "Display name",
                    "type": "string",
                    "maxLength": 100,
                    "x-nullable": true
                }
            }
        },
...
        "AssignedGroupSummary": {
            "title": "Assigned group",
            "required": [
                "name"
            ],
            "type": "object",
            "properties": {
                "id": {
                    "title": "ID",
                    "type": "integer",
                    "readOnly": true
                },
                "name": {
                    "title": "Name",
                    "type": "string",
                    "maxLength": 100,
                    "minLength": 1
                },
                "cover": {
                    "title": "Cover",
                    "type": "string",
                    "readOnly": true,
                    "format": "uri"
                },
                "isAdmin": {
                    "title": "Is admin",
                    "type": "string",
                    "readOnly": true
                }
            }
        },
        "Image": {
            "type": "object",
            "properties": {
                "id": {
                    "title": "ID",
                    "type": "integer",
                    "readOnly": true
                },
                "imgUrl": {
                    "title": "Img url",
                    "type": "string",
                    "readOnly": true,
                    "format": "uri"
                }
            }
        },
...
        "Post": {
            "required": [
                "content",
                "assignedUser",
                "assignedGroup",
                "photos"
            ],
            "type": "object",
            "properties": {
                "id": {
                    "title": "ID",
                    "type": "integer",
                    "readOnly": true
                },
                "content": {
                    "title": "Content",
                    "type": "string",
                    "maxLength": 3000,
                    "minLength": 1
                },
                "assignedUser": {
                    "$ref": "#/definitions/AssignedUserSummary"
                },
                "assignedGroup": {
                    "$ref": "#/definitions/AssignedGroupSummary"
                },
                "comments": {
                    "title": "Comments",
                    "type": "string",
                    "readOnly": true
                },
                "time": {
                    "title": "Time",
                    "type": "string",
                    "format": "date-time",
                    "readOnly": true
                },
                "updateTime": {
                    "title": "Update time",
                    "type": "string",
                    "format": "date-time",
                    "readOnly": true
                },
                "type": {
                    "title": "Type",
                    "type": "integer",
                    "enum": [
                        0,
                        1
                    ]
                },
                "photos": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Image"
                    }
                },
                "reactions": {
                    "title": "Reactions",
                    "type": "string",
                    "readOnly": true
                },
                "polls": {
                    "title": "Polls",
                    "type": "string",
                    "readOnly": true
                }
            }
        },
...
    }
}